+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Add `RequestPayload`, `RequestPath`, `RequestMethod`, `ResponseStatusCode` telemetry base property to `OCSDKContract`
+- Add `InvalidAuthentication` to `ExceptionThrower`
 
 ### Changed
 - Use `parseLowerCaseString()` on chat config properties to protect text case change